Enter foods can now delete an item that the user has entered by pressing the item's delete button. (The picture of the delete button may have to be messed with) Enter foods also can search by category (minimal functionality - still need to be able to go back up to category layer from food layer... Ex. Apple, Bananas -> back up to Fruit, Veggies...). Some visual upgrades to Enter Foods, top portion needs to be worked on.
Added a couple functions to description database to facilitate browsing by category.

Uploaded a test icon for the app. It is very much up for review and welcome to be changed. For now at least it is better than the default icon. 

Cleaned up some old files that aren't being used anymore. 

**NOTE: Updated the database so you will need to wipe your old ones. (We don't want to udate the version numbers constantly before the app is even done.) The description database now has the expiration dates in days according to soonest possible expiration. (e.g. 2 weeks = 14 days, 2-3 weeks = 14 days)